Move simulation-page/ to common/
