+++ components/common/index.ts
@@ -2,5 +2,6 @@ export { default as AppHeader } from "./app-header";
 export * from "./articles-inputs";
 export * from "./card"
 export * from "./expandable-panels";
+export { default as SimulationPage } from "./simulation-page";
 export * from "./tooltips";
 export * from "./utils";
