+++ components/common/simulation-page/SimulationPage.module.scss
