+++ components/common/simulation-page/SimulationPage.tsx
@@ -6,8 +6,8 @@ import { PureComponent } from "react";
 import SwipeableViews from "react-swipeable-views";
 
 // eslint-disable-next-line no-unused-vars
-import { ParametersState } from "../../redux/reducers/parameters";
-import { AppHeader } from "../common";
+import { ParametersState } from "../../../redux/reducers/parameters";
+import AppHeader from "../app-header";
 import { ArticleHeader } from "./article-header";
 import { SimulationMenuBar } from "./simulation-menu";
 import styles from "./SimulationPage.module.scss";
