+++ components/common/simulation-page/article-header/ArticleHeader.module.scss
