+++ components/common/simulation-page/article-header/index.ts
