+++ components/common/simulation-page/simulation-menu/button/Button.module.scss
