+++ components/common/simulation-page/simulation-menu/legende/Legende.module.scss
