+++ components/common/simulation-page/simulation-menu/legende/index.ts
